server now better equipped for async operations
