+++ sniper-common/src/lib.rs
@@ -29,10 +29,10 @@ pub trait SniperService {
     //async fn target_drop_libs(session_id: String, uri: String, libs: Vec<String>);
     
     /// get the triggers for the snippets associated with a given target
-    async fn get_triggers(session_id: String, uri: String)-> Vec<String>;
+    async fn get_triggers(session_id: String, uri: String) -> Vec<String>;
     
     /// get a snippet
-    async fn get_snippet(language: String, snippet_key: String) -> Vec<String>;
+    async fn get_snippet(language: String, snippet_key: String) -> Option<Vec<String>>;
     
 }
